Moved thumbnail stuff to helper class
